+++ overlays/haskell-nix-extra/utils/default.nix
@@ -1,37 +1,20 @@
-{ pkgs
-}:
+{ lib, haskell-nix, symlinkJoin }:
 
-with pkgs;
-
-rec {
-  # The utils derivation.
-  package = buildPackages.haskellPackages.callCabal2nix "iohk-nix-utils" ./. {};
-
-  # A function for building a script that can run in CI.
-  # The given script path, which is Haskell file will be able to
-  # import the CommonBuild module.
-  stackRebuild =
-   { script           # path to .hs file for building
-   , buildTools ? []  # extra programs to add to PATH
-   , libs ? (ps: [])  # selector function for extra haskell libraries
-   , shell ? null     # shell derivation to keep gc-rooted during build
-   }:
-   let
-      buildTools' = [
-        gnused coreutils git nix gnumake gnutar gzip lz4
-        stack stack-hpc-coveralls
-        haskellPackages.weeder
-      ] ++ buildTools;
-      libs' = ps: libs ps ++ [ package ];
-      ghc' = haskellPackages.ghcWithPackages libs';
-
-   in runCommand "stack-rebuild" {
-        buildInputs = [ ghc' makeWrapper ];
-    } ''
-      mkdir -p $out/bin
-      ghc -Wall -threaded -o $out/bin/rebuild ${script}
-      wrapProgram $out/bin/rebuild \
-        --set PATH "${lib.makeBinPath buildTools'}" \
-        --set NO_GC_STACK_SHELL "${toString shell}"
-    '';
-}
+let
+  project = haskell-nix.stackProject {
+    src = haskell-nix.cleanSourceHaskell {
+      name = "iohk-nix-utils-src";
+      src = ./.;
+    };
+  };
+in
+  symlinkJoin {
+    name = "iohk-nix-utils";
+    paths = lib.attrValues project.iohk-nix-utils.components.exes;
+    passthru = {
+      shell = project.shellFor {};
+      roots = project.roots;
+      package = builtins.trace "WARNING: iohk-nix `haskellBuildUtils.package` has been renamed to `haskellBuildUtils`." null;
+      stackRebuild = builtins.trace "WARNING: iohk-nix stackRebuild script has been removed." null;
+    };
+  }